enable map settings
